% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{euCodedFileHeader}
\alias{euCodedFileHeader}
\title{Generate List With EU Header Information}
\usage{
euCodedFileHeader(
  separator = ";",
  decimal = ".",
  quote = "\\"",
  encoding = "ISO-8859-1",
  language = "en",
  year = 2010
)
}
\arguments{
\item{separator}{default: ";"}

\item{decimal}{default: "."}

\item{quote}{default: '"'}

\item{encoding}{default: "ISO-8859-1"}

\item{language}{default: "en"}

\item{year}{default: 2010}
}
\value{
list with elements \code{separator}, \code{decimal}, \code{quote},
  \code{encoding}, \code{language}, \code{year}
}
\description{
Generate List With EU Header Information
}
