% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractObservationData.R
\name{extractObservationData}
\alias{extractObservationData}
\title{Extract Observations from EN13508.2-coded file}
\usage{
extractObservationData(euLines, headerInfo, header.info, file = "")
}
\arguments{
\item{euLines}{text lines read from EN13508.2-coded file}

\item{headerInfo}{data frame with information about header lines}

\item{header.info}{list as returned by
\code{kwb.en13508.2:::getFileHeaderFromEuLines}}

\item{file}{optional. Name of the file from which \code{euLines} were read.}
}
\value{
data frame with columns \code{A}, \code{B}, \code{C}, ... as defined 
  in EN13508.2 and a column \code{inspno} referring to the inspection number.
}
\description{
Extract Observations from EN13508.2-coded file
}
