% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getCodes}
\alias{getCodes}
\title{Get EU Codes and Their Meaning}
\usage{
getCodes(table = NULL, fields = NULL)
}
\arguments{
\item{table}{set to a name that is related to a specific table in the EU
norm. To get the possible table names, try \code{unique(getCodes()$Table)}}

\item{fields}{set to a vector of field (column) names to restrict the columns
returned}
}
\value{
data frame
}
\description{
Get a data frame containing EU codes and their meaning in different languages
}
