% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEuFormat.R
\name{getHeaderLinesFromHeaderInfo}
\alias{getHeaderLinesFromHeaderInfo}
\title{Get Header Lines From Header Info}
\usage{
getHeaderLinesFromHeaderInfo(header.info)
}
\arguments{
\item{header.info}{list with elements \code{encoding}, \code{language},
\code{separator}, \code{decimal}, \code{quote}, \code{year}}
}
\description{
Get Header Lines From Header Info
}
