% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLineDamageInfo.R
\name{getLineDamageInfo}
\alias{getLineDamageInfo}
\title{Get Information on Line Damages}
\usage{
getLineDamageInfo(observations, dbg = TRUE)
}
\arguments{
\item{observations}{data frame with observations. Required columns: \code{I}
(= horizontal or vertical position), \code{J} (= code for line damage),
\code{inspno} (number of inspection to which the observation belongs)}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
data frame with columns \code{ino} (inspection number), \code{ldno}
  (line damage number), \code{beg.at}, \code{end.at}, \code{beg.x} (position
  of line damage begin), \code{end.x} (position of line damage end),
  \code{length} (length of line damage)
}
\description{
Get Information on Line Damages
}
