% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndMergeEuCodedFiles.R
\name{mergeInspectionData}
\alias{mergeInspectionData}
\title{Merge Inspection Data}
\usage{
mergeInspectionData(x, warn = FALSE, naToEmpty = TRUE)
}
\arguments{
\item{x}{list of elements each of which represents inspection data read from
an EN13508.2-encoded file by means of \code{\link{readEuCodedFile}}.}

\item{warn}{logical indicating whether to warn about different header
information. By default, warnings are not shown.}

\item{naToEmpty}{logical indicating whether or not to replace \code{NA} with 
an empty string constant \code{""} in all columns of type character. The
default is \code{TRUE}.}
}
\value{
list with elements \code{header.info}, \code{inspections}, 
  \code{observations}.
}
\description{
Merge inspections and observations provided in a list
}
