% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotObservations}
\alias{plotObservations}
\title{Plot Observations per Pipe}
\usage{
plotObservations(survey, to_pdf = TRUE, matrix_dim = c(3, 2))
}
\arguments{
\item{survey}{list with elements \code{inspections} and \code{observations}
as e.g. returned by \code{kwb.en13508.2:::readEuCodedFile}}

\item{to_pdf}{if \code{TRUE} (default) the output goes into a temporary PDF
file}

\item{matrix_dim}{vector of two specifying the number of rows and columns,
respectivly of the matrix in which to arrange the single plots.}
}
\description{
Plot Observations per Pipe
}
\examples{
# Install packages if not yet installed
\dontrun{
install.packages("ggplot2")
devtools::install_github("guiastrennec/ggplus")
}

# Load example data
file <- system.file("extdata/example_13508_2.txt", package = "kwb.en13508.2")
survey <- kwb.en13508.2::readEuCodedFile(file)

# Create one plot per inspection in "survey"
kwb.en13508.2::plotObservations(survey, to_pdf = FALSE)

}
