% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobalInspectionID.R
\name{replaceInspectionId}
\alias{replaceInspectionId}
\title{Replace values in columns inspection_id}
\usage{
replaceInspectionId(inspection.data, new_ids)
}
\arguments{
\item{inspection.data}{list with inspections data frame in element 
\code{inspections} and observations data frame in element 
\code{observations}. Both data frames must have a column 
\code{inspection_id}.}

\item{new_ids}{vector of as many inspection ids as there are rows in 
\code{inspection.data$inspections} to be given to the inspections in that
data frame. The first element is given to the first row, the second to the
second row, and so on.}
}
\value{
list with data frames in elements \code{inspections} and 
  \code{observations}. In each data frame the values in column 
  \code{inspection_id} are updated according to the \code{new_ids}.
}
\description{
Replace values in columns inspection_id
}
\examples{
inspection.data <- list(
  inspections = data.frame(
    inspection_id = 1:3,
    pipe_id = 1:3
  ),
  observations = data.frame(
    inspection_id = c(1, 1, 2, 2, 3, 3),
    observation = c("start", "end", "start", "end", "start", "end")
  )
)
replaceInspectionId(inspection.data, new_ids = paste0("id_", 1:3))

}
