% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobalInspectionID.R
\name{setGlobalInspectionID}
\alias{setGlobalInspectionID}
\title{Set Global Inspection ID}
\usage{
setGlobalInspectionID(inspection.data, project = NULL, default.time = "22:33")
}
\arguments{
\item{inspection.data}{list with elements \code{header.info},
\code{inspections}, \code{observations}}

\item{project}{name of project to which the data are related, such as:
"Lausanne"}

\item{default.time}{default time string to use if column InspTime is not
available. Default: "22:33". A random number will be generated for the 
seconds, just to increase the chance that setting the time is enough to
generate a unique key.}
}
\value{
list with the same elements as in \code{inspection.data} but with
  columns \code{inspid} being added to the data frames "inspections" and
  "observations"
}
\description{
Convert inspections numbers (inspno) 1,2,3,... to globally unique identifiers
(inspid), such as "e4b48d86"
}
