% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobalInspectionID.R
\name{setInspectionId}
\alias{setInspectionId}
\title{Add column inspection_id to table of inspections and observations}
\usage{
setInspectionId(inspections, observations, inspection_ids)
}
\arguments{
\item{inspections}{data frame where each row represents an inspection}

\item{observations}{data frame where each row represents an observation. The 
data frame must have a column "inspno" that refers to a row in the data 
frame \code{inspections}.}

\item{inspection_ids}{vector of as many inspection ids as there are rows in 
\code{inspections}}
}
\value{
list with elements \code{inspections} and \code{observations} each of
  which has a new column "inspection_id" as its first column.
}
\description{
Add column inspection_id to table of inspections and observations
}
\examples{
inspections <- data.frame(pipe_id = 1:3)
observations <- data.frame(
  inspno = c(1, 1, 2, 2, 3, 3),
   observation = c("start", "end", "start", "end", "start", "end")
)
setInspectionId(inspections, observations, paste0("id_", 1:3))

}
