% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEuFormat.R
\name{toEuFormat}
\alias{toEuFormat}
\title{Generate Lines in EU Export Format}
\usage{
toEuFormat(inspection.data, version = 3L, ..., dbg = TRUE)
}
\arguments{
\item{inspection.data}{inspection data as retrieved by e.g.
\code{\link{readEuCodedFile}}}

\item{version}{version of implementation. One of \code{c(1, 2, 3)}}

\item{\dots}{passed to \code{toEuFormat_v2}}

\item{dbg}{whether or not to show debug messages}
}
\description{
Generate Lines in EU Export Format
}
