% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEuFormat.R
\name{toEuFormat_v2}
\alias{toEuFormat_v2}
\title{Generate Lines in EU Export Format (v2)}
\usage{
toEuFormat_v2(inspection.data, mycsv, ..., dbg = TRUE)
}
\arguments{
\item{inspection.data}{inspection data as retrieved by e.g.
\code{\link{readEuCodedFile}}}

\item{mycsv}{logical. If TRUE "my" version of writing CSV is used (fast),
otherwise CSV is written by means of write.table (slow)}

\item{\dots}{further arguments passed to dataFrameContentToTextLines}

\item{dbg}{whether or not to show debug messages}
}
\description{
Generate lines in EU export format (version 2: faster than version 1)
}
