% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{toEuFormattedLines_v2_3}
\alias{toEuFormattedLines_v2_3}
\title{Generate Lines in EU Export Format (v2)}
\usage{
toEuFormattedLines_v2_3(inspection.data, mycsv, ...)
}
\arguments{
\item{inspection.data}{inspection data as retrieved by e.g.
\code{\link{readEuCodedFile}}}

\item{mycsv}{logical. If TRUE "my" version of writing CSV is used (fast),
otherwise CSV is written by means of write.table (slow)}

\item{\dots}{further arguments passed to dataFrameContentToTextLines}
}
\description{
Generate lines in EU export format (version 2: faster than version 1)
}
