% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{valuesToCsv}
\alias{valuesToCsv}
\title{Values to CSV}
\usage{
valuesToCsv(x, dec = ".", sep = ",", na = "", qchar = "\\"",
  qmethod = c("double", "escape")[1])
}
\arguments{
\item{x}{vector of values representing a row in a CSV file}

\item{dec}{decimal character}

\item{sep}{field separating character}

\item{na}{text to be used in case of NA values}

\item{qchar}{quoting character to be used to surround text fields containing
the field separator}

\item{qmethod}{method used to indicate that a quoting character within a
quoted text field is not the ending quote. Either "double" (double the
quote character) or "escape" (backslash in front of the quoting character).}
}
\description{
Values to CSV
}
