% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{writeEuCodedFile}
\alias{writeEuCodedFile}
\title{Write Inspection Data to File in EU Format}
\usage{
writeEuCodedFile(inspection.data, output.file, version = 3, dbg = TRUE, ...)
}
\arguments{
\item{inspection.data}{inspection data as retrieved by e.g.
\code{\link{readEuCodedFile}}: list with elements \code{header.info},
\code{inspections}, \code{observations}. List element \code{header.info} is
a list with at least the following elements: \code{separator} (column
separator), \code{decimal} (decimal character "." or ","), \code{quote}
(character used to quote texts in order to allow the separator sign to be
used within the text). List element \code{observations} is a data.frame
with required columns \code{inspno} (inspection number)}

\item{output.file}{full path to output file}

\item{version}{version of implementation. One of \code{c(1, 2, 3)}}

\item{dbg}{if \code{TRUE} debug messages are shown}

\item{\dots}{passed to \code{toEuFormattedLines_v2_3}}
}
\description{
Write Inspection Data to File in EU Format
}
