% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{writeEuCodedFiles}
\alias{writeEuCodedFiles}
\title{Write Inspection Data to Files in EU Format}
\usage{
writeEuCodedFiles(survey, file, blocksize = 100, dbg = TRUE)
}
\arguments{
\item{survey}{list with elements \code{header.info}, \code{inspections}, 
\code{observations}, just as required for parameter "inspection.data" in 
\code{\link{writeEuCodedFile}}.}

\item{file}{Full path to output file. The file name must end with ".txt"
which will be replaced with "_<i>_<j>.txt" with i and j being the number of
the first and last inspection, respectively, contained in the file.}

\item{blocksize}{number of inspections to be written to one file. Default:
100}

\item{dbg}{if TRUE (default) debug messages are shown.}
}
\description{
Write inspection data to files in EU format with each file containing data of
a fix number (default: 100) of inspections
}
\seealso{
\code{\link{writeEuCodedFile}}
}
