% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractObservationData.R
\name{extractObservationData}
\alias{extractObservationData}
\title{Extract Observations from EN13508.2-coded file}
\usage{
extractObservationData(
  euLines,
  headerInfo,
  header.info,
  file = "",
  as.text = FALSE
)
}
\arguments{
\item{euLines}{text lines read from EN13508.2-coded file}

\item{headerInfo}{data frame with information about header lines}

\item{header.info}{list as returned by
\code{kwb.en13508.2:::getFileHeaderFromEuLines}}

\item{file}{optional. Name of the file from which \code{euLines} were read.}

\item{as.text}{whether or not to keep columns in their original (character) 
type. The default is \code{FALSE}, i.e. columns that are expected to
contain numeric values are converted to numeric, respecting the decimal
separator that is given in \code{header.info}}
}
\value{
data frame with columns \code{A}, \code{B}, \code{C}, ... as defined 
  in EN13508.2 and a column \code{inspno} referring to the inspection number.
}
\description{
Extract Observations from EN13508.2-coded file
}
