% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodes.R
\name{getCodes}
\alias{getCodes}
\title{Get EU Codes and Their Meaning}
\usage{
getCodes(table = NULL, fields = NULL)
}
\arguments{
\item{table}{name or vector of names of tables in the EU norm for which to 
get field information. Use \code{unique(getCodes()$Table)} to get the 
possible table names.}

\item{fields}{set to a vector of field (column) names to restrict the columns
returned}
}
\value{
data frame
}
\description{
Get a data frame containing EU codes and their meaning in different languages
}
