% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeInspectionData}
\alias{mergeInspectionData}
\title{Merge Inspection Data}
\usage{
mergeInspectionData(inspectionDataList)
}
\arguments{
\item{inspectionDataList}{list of elements each of which represents
inspection data read from an EN13508.2-encoded file by means of
\code{\link{readEuCodedFile}}.}
}
\value{
list with elements \code{header.info}, \code{inspections}, 
  \code{observations}.
}
\description{
Merge inspections and observations provided in a list
}
