% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{readEuCodedFiles}
\alias{readEuCodedFiles}
\title{Read Multiple CCTV Inspection Files}
\usage{
readEuCodedFiles(input.files, dbg = TRUE, append.file.names = TRUE,
  ...)
}
\arguments{
\item{input.files}{vector of character paths to input files}

\item{dbg}{if \code{TRUE} debug messages are shown}

\item{append.file.names}{if TRUE (default) the filename will be provided in
column \code{file} in the \code{inspections} element of the result list}

\item{\dots}{arguments passed to \code{\link{readEuCodedFile}}, such as 
\code{read.inspections}, \code{simple.algorithm}, \code{warn}, see there.}
}
\value{
list of sublists each of which is the result of a call to 
  \code{\link{readEuCodedFile}}. The names of the list elements are the file 
  names of the input files, with a preceding letter "x" and special 
  characters replaces with underscore.
}
\description{
Read CCTV inspection data from multiple files coded according to EN13508-2
}
