% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobalInspectionID.R
\name{setGlobalInspectionID}
\alias{setGlobalInspectionID}
\title{Set Global Inspection ID}
\usage{
setGlobalInspectionID(inspection.data, project = NULL)
}
\arguments{
\item{inspection.data}{list with elements \code{header.info},
\code{inspections}, \code{observations}}

\item{project}{name of project to which the data are related, such as:
"Lausanne"}
}
\value{
list with the same elements as in \code{inspection.data} but with
  columns \code{inspid} being added to the data frames "inspections" and
  "observations"
}
\description{
Convert inspections numbers (inspno) 1,2,3,... to globally unique identifiers
(inspid), such as "e4b48d86"
}
